Added addStudentToCourse and withdrawStudentFromCourse
Database connection is somehow closed after DELETE operations. This may be done automatically I am not sure. Weird thing is I only get the database is closed error with UPDATE, it did not occur when I added a new course, an INSERT operation. I also ran into this issue after the boot() method a few days ago. It was saying the database was closed after booting and trying to add a new course so I put a isClosed() check after boot. We may need to check if the connection is closed before each method but it seems to work fine for now. This is the check if we add it later it is already done in updateSchedule() method:
 try {
            if(databaseConnection.isClosed()) {
                databaseConnection = DriverManager.getConnection("jdbc:sqlite:sisao.db");

            }
        } catch (SQLException e) {
            e.printStackTrace();
        }